requirements.txt
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.1
scipy==1.12.0
torch==2.3.0
torch_scatter==2.1.2
sentence-transformers==2.7.0
Unidecode==1.3.8
python-terrier==0.10.1
nltk==3.8.1
PyYAML==6.0.1
pip==24.0
setuptools==69.1.1
statsmodels==0.14.1
transformers==4.38.2